+++ suslab/users/controllers.py
@@ -1,9 +1,6 @@
-from flask import request, render_template, url_for, redirect
 from flask_security import SQLAlchemyUserDatastore
 
 
 def get_user_datastore():
-    from suslab.users.models import User, Role
-    from suslab import db
-
+    from suslab.users.models import User, Role, db
     return SQLAlchemyUserDatastore(db, User, Role)
